application/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
gunicorn==21.2.0